+++ sdk/openai/openai/src/api/operations.ts
@@ -29,6 +29,7 @@ import {
   BeginAzureBatchImageGeneration202Response,
   BeginAzureBatchImageGenerationDefaultResponse,
   BeginAzureBatchImageGenerationLogicalResponse,
+  ChatMessage as GeneratedChatMessage,
   OpenAIContext as Client,
   GetChatCompletions200Response,
   GetChatCompletionsDefaultResponse,
@@ -241,7 +242,7 @@ export function _getChatCompletionsSend(
   return context.path("/deployments/{deploymentId}/chat/completions", deploymentId).post({
     ...operationOptionsToRequestParameters(options),
     body: {
-      messages: messages,
+      messages: parseChatMessage(messages),
       functions: options?.functions,
       function_call: options?.functionCall,
       max_tokens: options?.maxTokens,
@@ -376,7 +377,7 @@ export function _getChatCompletionsWithAzureExtensionsSend(
     .post({
       ...operationOptionsToRequestParameters(options),
       body: {
-        messages: messages,
+        messages: parseChatMessage(messages),
         functions: options?.functions,
         function_call: options?.functionCall,
         max_tokens: options?.maxTokens,
@@ -785,4 +786,14 @@ export async function getAudioTranscription<Format extends AudioResultFormat>(
   return response_format !== "verbose_json"
     ? body
     : (renameKeysToCamelCase(body) as AudioResult<Format>);
+}
+
+function parseChatMessage(messages: ChatMessage[]): GeneratedChatMessage[] {
+  return messages.map((p: ChatMessage) => ({
+    role: p.role,
+    content: p.content ?? null,
+    name: p.name,
+    function_call: p.functionCall,
+    context: p.context,
+  }));
 }