+++ sdk/openai/openai/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Returns `usage` information when available.
 - Fixes a bug where errors weren't properly being thrown from the streaming methods.
 - Returns `error` information in `ContentFilterResults` when available.
+- Fixes parsing of `functionCall` in `ChatMessage` objects.
 
 ## 1.0.0-beta.5 (2023-08-25)